+++ Harbor_employee/src/main/java/com/example/harbor_employee/Employee/controller/EmployeeController.java
@@ -1,8 +1,8 @@
 package com.example.harbor_employee.Employee.controller;
 
 import com.example.harbor_employee.Employee.dto.request.EmployeeSearchDto;
-import com.example.harbor_employee.Employee.dto.response.EmployeeDetailResDto;
 import com.example.harbor_employee.Employee.dto.response.EmployeeResDto;
+import com.example.harbor_employee.client.dto.LoginMemberResDto;
 import com.example.harbor_employee.global.common.CommonResponse;
 import com.example.harbor_employee.Employee.service.EmployeeService;
 import lombok.RequiredArgsConstructor;
@@ -10,9 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -27,5 +25,11 @@ public class EmployeeController {
     public ResponseEntity<CommonResponse> getList(EmployeeSearchDto employeeSearchDto, Pageable pageable){
         List<EmployeeResDto> employees = employeeService.findAll(employeeSearchDto, pageable);
         return new ResponseEntity<>(new CommonResponse("요청이 정상적으로 실행되었습니다.", employees), HttpStatus.OK);
+    }
+
+    @PostMapping("/create")
+    public HttpStatus BasicCreateEmployee(@RequestBody LoginMemberResDto loginMemberResDto) {
+        log.info(" 호출 시작 : ");
+        return employeeService.createBasicEmployee(loginMemberResDto);
     }
 }
