--- Harbor_employee/src/main/java/com/example/harbor_employee/client/GetUsersResponse.java
@@ -1,10 +1,0 @@
-package com.example.harbor_employee.client;
-
-import lombok.Data;
-
-@Data
-public class GetUsersResponse {
-    String email;
-    String name;
-    String birth;
-}
