+++ Harbor_employee/src/main/java/com/example/harbor_employee/client/testClient.java
@@ -10,6 +10,4 @@ import org.springframework.web.bind.annotation.RequestParam;
 @FeignClient(name = "testClient", url = "http://localhost:8001")
 public interface testClient {
 
-    @GetMapping(value = "/login/detail")
-    GetUsersResponse getUsers(@RequestParam String email);
 }
