--- Harbor_employee/src/main/java/com/example/harbor_employee/filter/JwtAuthFilter.java
@@ -1,63 +1,0 @@
-package com.example.harbor_employee.filter;
-
-import com.example.harbor_employee.config.CustomUserDetails;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationServiceException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class JwtAuthFilter extends GenericFilter {
-
-    @Value("${jwt.secretKey}")
-    private String secretKey;
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        try {
-            String bearerToken = ((HttpServletRequest)request).getHeader("Authorization");
-            if(bearerToken != null) {
-                //bearer token에서 token값만 추출
-                if(!bearerToken.startsWith("Bearer ")) {
-                    throw new AuthenticationServiceException("token의 형식에 맞지 않습니다.");
-                }
-                String token= bearerToken.substring(7);
-
-                System.out.println("토큰 검사 실행");
-
-                Claims claims = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody();
-                String employeeId = claims.get("employeeId").toString();
-                //Authentication 객체를 생성하기 위한 UserDetails 생성
-                CustomUserDetails customUserDetails = CustomUserDetails.builder()
-                        .role(claims.get("role").toString())
-                        .username(claims.getSubject())
-                        .password("")
-                        .employeeId(employeeId)
-                        .build();
-
-                customUserDetails.getAuthorities().add(new SimpleGrantedAuthority("ROLE_" + claims.get("role")));
-
-                UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(
-                        customUserDetails, null, customUserDetails.getAuthorities()
-                );
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-            chain.doFilter(request, response);
-
-        } catch (Exception e) {
-            HttpServletResponse httpServletResponse = (HttpServletResponse) response;
-            httpServletResponse.setStatus(HttpStatus.UNAUTHORIZED.value());
-            httpServletResponse.setContentType("application/json");
-        }
-    }
-}
