+++ Harbor_employee/src/main/java/com/example/harbor_employee/global/support/Role.java
@@ -1,0 +1,6 @@
+package com.example.harbor_employee.global.support;
+
+public enum Role {
+    USER,ADMIN;
+}
+
