--- Harbor_employee/src/main/java/com/example/harbor_employee/support/Role.java
@@ -1,6 +1,0 @@
-package com.example.harbor_employee.support;
-
-public enum Role {
-    USER,ADMIN;
-}
-
